Initial commit
This app utilizes both UIKit and SwiftUI frameworks. UIKit is used to connect to a microcontroller via CoreBluetooth, continuously read the data sent from a peripheral, save the data via CoreData, and transfer that data to SwiftUI views. SwiftUI is used to graph the data via Charts and display statistics.